fix: don't compile code with annotations future flag
This is technically a breaking change, but it better matches the user
code.

Note that it is not currently possible to save a notebook file with
`from __future__ import annotations`, since Python requires future
imports to be at the top of the file. If people want deferred
annotations we can make it configurable.