fix: don't compile code with annotations future flag (#3918)
This changes our compiler to not inherit the main program's compiler
flags; in particular this means that user code is not compiled with
deferred/stringified annotations (`__future__.annotations`).

This is technically a breaking change, but it better matches user code.

Note that it is not currently possible to save a notebook file with
`from __future__ import annotations`, since Python requires future
imports to be at the top of the file. If people want deferred
annotations we can make it configurable.

Fixes #3915, #3312