fix: cell staleness when autorun on startup is off
Previously, when autorun on startup was disabled, running a cell
wouldn't transition it out of stale. For long-running cells, this would
give the incorrect impression to the user that their cell had not been
submitted for execution.

This change updates the kernel to register uninstantiated cells as
stale, so they are correctly transitioned out of stale on run.