fix: default to opt-in for introspection of tables (#3940)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
Not sure what you think about this, but this may be a temporary solution
for the issue here:
https://discord.com/channels/1059888774789730424/1338959377411084390/1344967482577915905.

By default, we do not introspect tables, we let users opt-in by changing
config.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->
For future MRs:
- [ ] ~remove introspection for meta schemas, make this a config.
`information_schema` does not need to be introspected I think..~ (if we
load on-demand, we do not need to do this)
- [ ] possible to request in the background(?) or same as table details
preview, request on-demand.

## 📋 Checklist

- [X] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [X] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick