experimental: reactive pytest (#3938)
## 📝 Summary

Follow up to: #3892

This allows eligible cells to automatically and reactively run pytest in
a notebook:


![image](https://github.com/user-attachments/assets/d5040319-7435-4ec4-bfa0-ca12f9b3f93b)

It also hooks into the runtime, so tests don't have to run from scratch

This is currently hidden under an experimental flag for a few reasons:

1. I noticed live editing the cells sometimes puts them in an invalid
state requiring another run (pytest loads the module off disk, and I
think there's a bit of a lag)
2. similar to top level functions, I think that it seems like a bit of
magic for when the functionality turns on. I think some more
documentation and UX communication might be needed before an intentional
release.
3. My unit tests aren't super robust and I haven't tested out every edge
case
4. Would nice to get some feedback before pushing too much further
 
Just a note, this is tightly integrated with pytest, I don't think there
are many other meaningful frameworks now nose is dead, but marimo should
attempt to be agnostic.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick @koaning

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>