fix: imports being stripped by vanilla ruff (#3952)
## 📝 Summary

A user on discord showed interest in playing around with top level
definitions and reported that none of their imports were preserved (and
hence their functions threw name errors when using imports).

I realized all my testing on the top level function stuff is from the
root of the marimo repository, which has F401 as "unfixable" in the
pytoml.

## 🔍 Description of Changes

Calls ruff with flag to prevent import removal.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>