fix: anywidget should re-render whenever re-created (#3998)
Fixes #3962

Previously, if the JS of an anywidget didn't change, we skipped a
re-render.

This could cause issues if there is initialization code in the render
function, because it would not get called when the widget is created
again.

Instead we not just re-render the anywidget whenever it gets created.