improvement: better prompt-to-install when missing deps, grouped installed, Optional Features panel (#4080)
Fixes #3113 

<img width="1293" alt="Screenshot 2025-03-12 at 9 30 17 PM"
src="https://github.com/user-attachments/assets/dc817d7a-7430-49d4-a9fa-3e0378c4c594"
/>

* Remove 'SQL' from KernelCapabilities. If the user tries to run SQL, we
will not prompt to install the necessary deps.
* We can now prompt the user for multiple missing deps
* Update `ai` endpoints to throw missing modules so that we can prompt
the user as well.
* `Optional Dependencies` panel in the settings menu to show the enabled
features and install. This is both for education and ease of
installation.

I didn't want to put everything in `KernelCapabilities` since its
another source of truth, and requires checking it on startup, which
could be a waste of resources/blocking