experimental: lsp server (#3414)
This PRs adds propper LSP integration, using `python-lsp-server` which
relies on `jedi`

Related to #1163, but does not support pyright

There are still outstanding issues before even showing the experimental
flag in the `Labs` settings, but going to merge this as-is in order to
prevent more merge conflicts/drift from main.

- [x] what is the correct LSP server? → `pylsp`
- [ ] add to `pyproject.toml` deps (`marimo[recommended]` or
`marimo[lsp]`) (what plugins should we include in recommended)
- [ ] compare with Jedi output
- [ ] docs
- [ ] add to `Labs` UI
- [x] include all cells in the document
- [x] linting
- [x] remove private methods (or put at the bottom)
- [x] `import marimo as mo; mo.` does not auto-complete
- [ ] custom `pylsp_completions` hook to use custom google doc string
formatting.
- [x] diagnostics (i.e. lint errors) update all cells, and not just the
currently focused cell

---------

Co-authored-by: Michał Krassowski <5832902+krassowski@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>