fix: Enable coverage of _save tests (#4111)
@mscolnick re https://github.com/marimo-team/marimo/pull/4038 tests only
seem to be failing in coverage. Seems to be because `with_block` does
funny things with sys.trace which coverage uses. This mitigates it a
bit- but seems to be a bit strange with just the first test which seems
to have a messed up code pointer (which may well be a python bug? Seems
to always fail on whatever the first test is)