improve: indication when cell output is loading (#4155)
Decrease opacity of stale loading outputs to make staleness more easily
parsed visually.

When cells are queued or running (not edited), it needs to be clear that
their output is loading and stale. This is important both for editing
notebooks and running long-running cells that have big outputs, as well
as for app view.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>