feat: add scroll to cell (#4162)
Closes #3346

Add a new cell action to copy a link to it, and logic to scroll to the
cell on startup.

This does require naming the cell. We could maybe later add using the
cell-id, but that does change when the order changes.

<img width="1137" alt="Screenshot 2025-03-19 at 11 47 05 AM"
src="https://github.com/user-attachments/assets/27ba97a3-61bb-4462-956b-d5e69d2967e7"
/>