fix: getting stuck waiting for socket to connect (#4213)
## 📝 Summary

Fixes one of the potential causes for #4067 

## 🔍 Description of Changes

Marimo was getting completely stuck for me, waiting for a timeout that
[will never
happen](https://docs.python.org/3/library/socket.html#notes-on-socket-timeouts).

(At least for WSL - I have no idea if it worked in other OS, but if it
did, I'm confused as for why that was not a problem)

## 📋 Checklist

- [X] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [X] I have run the code and verified that it works as expected.

## 📜 Reviewers

@mscolnick suggested looking into that function in the linked issue