Styling cells (#4245)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
Hello, as asked on Discord, we are hoping to add some styling options to
the cells of a data-table.


![image](https://github.com/user-attachments/assets/c172d02b-32bd-450e-8154-0d5276aff939)

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

On the Python side, I'm adding a new function that returns a
dictionary-style object to describe the cell's style. This nested array
of styles is then sent to the client and utilized by another TanStack
plugin.

In our Discord conversation, we discussed foreground and background
colors. I opted for a style dictionary, which provides more options. I'm
not set on this idea, so please share your thoughts on what you think is
best.

This is still a work in progress, so I would appreciate your guidance in
the right direction.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@mscolnick

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>