improve: override sys.argv in kernel
This change sets sys.argv to match the CLI extra arguments, i.e.
the arguments after the filename, in edit and run sessions, to match
semantics when running as a script.

This makes it possible to write a single argparse or simple-parsing
specification that works when:

1  editing or running a notebook as an app
2. running as a script
3. exporting to HTML / IPYNB

For example, for

marimo edit notebook.py -- foo bar -baz 42

sys.argv would be set to

["notebook.py", "foo", "bar", "-baz", "42"]

With this change, mo.cli_args() is not really needed anymore.