docs: cli arguments (#4304)
Improve the docs on how to run marimo notebooks as scripts with CLI
arguments, using argparse and `simple-parsing`.

This documentation intentionally omits Typer, because Typer leads to
notebooks whose execution branches depending on whether running as a
script or as a notebook. The included examples have the same control
flow for both execution modes (except for the argument parsing logic).