feat: argparse support when running as a notebook (#4306)
This change makes the kernel set `sys.argv` to the CLI extra arguments,
i.e. the arguments after the filename, in edit and run sessions and when
exporting. This matches semantics when running as a script.

This makes it possible to write an argparse or simple-parsing
specification that works when:

1.  editing or running a notebook as an app
2. running as a script
3. exporting to HTML / IPYNB

Without this change, argparse worked for 2 but not for 1 and 3.

For example, for

```
marimo edit notebook.py -- foo bar -baz 42
```
`sys.argv` would be set to

```
["notebook.py", "foo", "bar", "-baz", "42"]
```

With this change, `mo.cli_args()` is not really needed anymore, but we
can keep it around (it has some [public
usage](https://github.com/search?q=mo.cli_args&type=code)).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>