improve: don't unselect rows when clicking in other rows
This change modifies the table to persist selections when clicking
anywhere in the table.

Previously, clicking anywhere in a table unselected (in the case of
single selection) an existing selection, or added additional selections
(in the case of multi).

Even though other tables might do this, this can sometimes be
troublesome since selections and unselections trigger code execution
that may be expensive. It also makes it not possible to copy text when a
row is already selected. One could try detecting for drag, but that
doesn't account for double click, and I think this issue
would still come up.

In favor of being explicit over implicit, and prioritizing edit mode
usage, it makes sense to have selection toggles be limited to
the column checkboxes.