fix: don't render hstack/vstack in extra divs if there is no flex config (#4393)
This could be a breaking change for some formats. The only cases I can
think of:
- people were expecting a wrapping div and created custom css to target
this

Currently if no `widths/heights` setting is passed to hstack/stack, then
we wrap them in `<div>`s. This can make it hard to style the child
element's width/height because the wrapping `div`.

This doesn't change when the output `widths/heights` are passed in.