fix: place nulls at last (#4436)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
Fixes #4381. For narwhals, pandas, polars, the change is
straightforward. But it's trickier for default table.

Use tuple is None technique for default table sorting:
[stackoverflow](https://stackoverflow.com/questions/18411560/sort-list-while-pushing-none-values-to-the-end)

There is one issue:
Sorting has never worked on dictionary (row-oriented data)

![image](https://github.com/user-attachments/assets/6c49016c-01a3-4635-b119-75b964d9637d)

Because it's column names are not 'key' and 'value', so I've made a
change to fix this. Not sure if it may lead to other bugs.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [X] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [X] I have added tests for the changes made.
- [X] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick