improve: vimrc support (#4487)
## 📝 Summary

Change the way vimrc gets applied
It is not an improvement of the logic but more of a rewrite
I referenced the vim manual while writing this
Some behavior is still not the same as vim but to me it seems good
enough at this point

## 🔍 Description of Changes

- Extract vimrc parser into a separate file
- Change parser implementation
- Rewrite tests to the new implementation
- Use correct methods for applying parsed vim commands

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@mscolnick

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>