security: statically load marimo notebooks for run and edit (#4400)
## 📝 Summary

Tests finally pass :tada:

This adds static loading, that is opposed to loading a notebook as a
module, the file is parsed directly and a notebook is constructed.

---

@Haleshot don't make this a priority but if you wanted to take the time
to do a code review on, here's your chance! A couple thoughts when doing
a code review:

## Docs: Do I understand what's going on?
I think the most key thing is that code is meant to be read by both
computers and _humans_.
If something is exceptionally unclear, the issue is likely in the code
and not you.
Doc strings are key- but 

## Is the code maintainable / brittle?

## Are the tests sufficient?
Ideally everyhting should be covered

## PR structure: Is the PR limited in scope?
Feature creep definitely happens in PRs, but ideally they should be
relatively contained.
Since marimo squashes commits in PRs, some PRs (like this one initially,
but I rebased for readability) can contain superfluous commits. Ideally
every commit should contain a little piece of the story- which makes
reviewing by examining the commits much easier (and ideal)

---


This is a far from perfect PR, so good practice (:

---------

Co-authored-by: Akshay Agrawal <akshay@marimo.io>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>