fix: pass setup cell to toplevel resolution posthook (#4538)
## 📝 Summary

Fixes some of this issues I was seeing when writing the tutorial. Main
issue seemed to be unable to discover the setup cell

Note we need actual cell order to properly do resolution, but that i
information is not available runner level. Could wire that down, but
this is OK for now.

Really needs some e2e tests to properly capture, but the logic in
toplevel itself is throughly tested in `tests/_ast/test_toplevel.py`

@akshayka OR @mscolnick