docs: changelog (#4592)
Link to the GitHub releases for the changelog, using it as the source of
truth. (This is what Polars does too.)

In the future we can mirror the release notes in our docs on a dedicated
page. But I didn't want to invest the effort in that right now, and this
is still an improvement.