feat: extend support to tqdm update() (#4623)
## 📝 Summary


Fixes #4622
Expands support to tqdm by adding update attribute to the marimo patched
tqdm

## 🔍 Description of Changes

Added update attribute to the tqdm patch so we can use huggingface
libraries with marimo and have a nice progressbar rendering.
Video of the solution working:

https://github.com/user-attachments/assets/b0aed94f-f520-472e-930e-419a502c6069

Video of the current version and the error when using update:

https://github.com/user-attachments/assets/5e6925d3-c74e-403d-8e2c-bdafffa7f62c


## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick