feat: add sql formatting button to sql cells (#4640)
Adds SQL formatting to the footer of the SQL cell.

This is currently only used by explicitly clicking the format button.
We do not use it for auto-formatting onSave or globally because
SQL formatting is much more opinionated than Python formatting, and we
don't want to tie the two together (just yet).



![image](https://github.com/user-attachments/assets/47c2ce3f-f58a-4c9f-a454-a6b44690cc85)

![image](https://github.com/user-attachments/assets/53cd7a40-eb65-402d-b474-67d0b2063e3c)