improve: scratchpad hint on multiple definition error (#4647)
The scratchpad is very handy when working with throwaway
code. The time to use it is often when you hit a multiple
definition error, so this change adds a suggestion/button
to open the scratchpad from the multiple definition error output.

This change was informed by watching students use marimo for the first
time; they loved the scratchpad panel for this use case when shown how
to open it.

<img width="793" alt="image"
src="https://github.com/user-attachments/assets/e39987ed-0f7e-4e9d-8304-41488c1a0b9b"
/>

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>