feat: Ibis SQL engine (#4683)
## 📝 Summary

Implements #4546 

A `IbisEngine` class was added, mimicking the `SQLAlchemyEngine`. It
automatically pickups Ibis `Backend` objects and displays them in the
database panel.

## 🔍 Description of Changes

The code reimplements the logic of the `SQLAlchemyEngine`. It was done
by implementing the `IbisEngine` and adding code wherever
`SQLAlchemyEngine` was referenced.

 ### Notes
- Ibis doesn't use the hierarchy `table <- schema <- database`. Instead,
it uses `table <- database <- catalog`
- Could add a `nested` / `complex` / `object` type in the UI. Right now,
everything defaults to `string`
- Could add a `Add table to notebook` in Python instead of SQL. It would
insert the statement
   ```python
   table = con.table("table_name")
   ```
- More thorough testing could be done to verify individual Ibis
backends.
- Could add a frontend component with a `Ibis` button that adds code to
the cell (similar to SQLAlchemy)


![image](https://github.com/user-attachments/assets/3e1b23bc-2e0a-4035-8de4-0bf13c9358cf)

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

I chatted with @mscolnick and see that @Light2Dark assigned the issue
two days ago

---------

Co-authored-by: zilto <zilto@github.com>
Co-authored-by: Myles Scolnick <myles@marimo.io>