feat: support ipython display update (#4772)
## 📝 Summary

Current iPython Display implementation doesn't support updates, it keeps
appending to the cell output
This is how it looks today when using HugginFace transformers in
notebook mode:


https://github.com/user-attachments/assets/b4c9bb17-9a39-4488-8996-9a6f07d9bae7




## 🔍 Description of Changes

This PR expands the support to Python display so it now returns a
DisplayHandle and that DisplyHanle has an update function.
This is how the change works when using HuggingFaace Transformers:


https://github.com/user-attachments/assets/a5819222-7cae-401c-bb2c-1f24489df120



## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

<!--
Tag potential reviewers from the community or maintainers who might be
interested in reviewing this pull request.

Your PR will be reviewed more quickly if you can figure out the right
person to tag with @ -->

@akshayka OR @mscolnick