improvement: use script metadata for config and deps in pyodide (#4806)
This adds support for inline-script metadata for pyodide.

If there are dependencies declared inline, then we use those. Otherwise
we fallback to the pyodide parsing of imports (which was what we did
previously).

This also parses inline script metadata for config.

--------------


This looks like more lines of code, but mostly:
1. lots of tests written and backfilled
2. some utils moved around