feat: mo.thread lifecycle (#4782)
Closes #4564 

This change adds a `should_exit` property to `mo.Thread` which evaluates
to `True`
when the thread's spawning cell lifecycle has ended. This makes it
possible for developers to clean up their threads.

This change also adds a utility method, mo.current_thread(), that
returns the current thread of control if it is a marimo.Thread object,
and raises otherwise. Developers could technically use
threading.current_thread() instead, but this method provides
type safety and validation.

---------

Co-authored-by: Dylan Madisetti <dylan@marimo.io>