CI build improvements (#4913)
## 📝 Summary

- Adds linux-64 to pixi config (was mac only)
- Adds a GH Actions workflow to build frontend & backend using pixi
- Extends max memory allocation for frontend to 4 GB

## 🔍 Description of Changes

Pixi is a good idea for providing a more consistent development
environment. Adding linux to the pixi config worked right off the bat.
If the code ever gets to a point where mac and linux need different
build dependencies, that is configurable through pixi.

The Github Action required a pixi.lock file. I think it's a good idea to
add this anyway so that devs have a reproducible environment. Sometimes
lockfiles require a little extra attention to git workflows and
maintenance to avoid merge conflicts, but it's worth the tradeoff in my
experience (usually just stashing or deleting them when pulling changes
works fine).

All of the above paragraph also applies for the pnpm lockfile which was
also added.

The mac image hit an out-of-memory error; this is also an error i have
consistently run into on my local linux environment FWIW. I extended the
max memory allocated to node by setting an environment variable in the
package.json build script configuration. Worked like a charm. Maybe the
frontend devs might want to see what the largest memory consumers are
here?

Regarding the new workflow: i'd be happy to merge the build into
existing workflows (such as test or release workflows) but figured i'd
get some feedback first; but in general i think a there's a lot of room
consolidation of development environment configurations both in github
actions and in the local configuration setup... happy to put in more
work on that with the go-ahead from the marimo team, and also happy to
put my thoughts down in more detail on a separate issue.


## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

## 📜 Reviewers

@dmadisetti @mscolnick